Move spell targeting over too
