+++ src/magic/spell_targeting.cpp
@@ -1,4 +1,4 @@
-#include "spell_targeting.h"
+#include "magic/spell_targeting.h"
 
 #include "character.h"
 #include "creature.h"
