+++ src/magic/spell_targeting.h
